Bot for bumping gix-components or ic-js (#5195)
# Motivation

When making changes in gix-components or ic-js, we always need to pull
in the changes into nns-dapp.
This is very straightforwards and is kind of pointless to review.
If those changes can be made by a bot, only one person needs to review
it so nobody needs to be blocked.

# Changes

Add a workflow which can be used to automatically update gix-components
and/or ic-js.

# Tests

This PR was
[created](https://github.com/dfinity/nns-dapp/actions/runs/9938705200/job/27451714562)
with it: https://github.com/dfinity/nns-dapp/pull/5194
I just changed the PR title after doing that test.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary