Delete Reproducible Assets workflow (#5215)
# Motivation

Early 2023 we tried to make the nns-dapp build reproducible across
platforms.
We succeeded in making the frontend assets reproducible but making the
Rust build reproducible across Linux and Mac does appear to currently be
possible.
So we've always been using Docker for reproducible builds.

We still have the GitHub workflow to test that our frontend assets are
reproducible.
But this is a maintenance burden and isn't useful without the Rust build
being reproducible.
So let's just delete that workflow. 

# Changes

Deleted .github/workflows/reproducible-assets.yaml

# Tests

deleted

# Todos

- [x] Add entry to changelog (if necessary).