Add getters on SnsSummaryWrapper for fields from swap.params (#5217)
# Motivation

We want to stop using the `summary.swap` field because it's deprecated
and everything on it is available in other ways.
One of the things we get from the `swap` field is the `params` object
which in turn has several fields.
In this PR we introduce a getters on `SnsSummaryWrapper` to get the
fields we care about from `swap.params`.
This will make it easy to switch over in another PR.

# Changes

1. Add `getSwapDueTimestampSeconds`, `getMinIcpE8s`, `getMaxIcpE8s`,
`getMinParticipants`, `getMinParticipantIcpE8s`,
`getMaxParticipantIcpE8s`, and `getSnsTokenE8s` to `SnsSummaryWrapper`
which returns the corresponding fields from `summary.swap.params`.
2. Use these getters instead of getting fields directly from
`summary.swap.params`.
3. Drive-by: Restore `// Use 1 instead of using enum to avoid importing
sns-js` comment which was accidentally removed in
https://github.com/dfinity/nns-dapp/pull/5192

# Tests

1. Existing unit tests updated.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary