NNS1-3139: Extract MaturityWithTooltip from NeuronMaturityCell (#5245)
# Motivation

We want to show maturity in the projects table, similar to the neurons
table, with a tooltip that shows separate available and staked maturity.
However there will be some different logic for when the value isn't
available so we can't reuse the full cell component.

# Changes

Extract `MaturityWithTooltip` component from `NeuronMaturityCell`.

# Tests

1. Unit test added.
2. Existing NeuronsTable tests pass.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary