Making e2e navigation less brittle (#5246)
# Motivation

There are 2 issues with the back button in end-to-end tests:
1. After clicking the back button, it may be replaced with the menu
button or there might still be a back button.
2. When navigating to another tab, we need to click on the menu button,
but this is not possible if there is a back button.

This PR solves this by making the `AppPo` aware of which views have a
back button.

# Changes

1. Add `getCurrentViewWithBackButton` method to `AppPo` which knows
which views have a back button and checks if any of them is currently
present to return it.
2. Remove the conditional option to wait for the back button to be
absent from `AppPo.goBack` and instead wait for the current view to be
absent.
3. Add a `goBackAllTheWay` method which can be used before navigating to
another tab without having to know if a back button first needs to be
clicked one or more times.
4. Remove calls to `goBack()` before navigating because that's now taken
care of by the navigation functions.

# Tests

Test pass.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary