Output canister logs during deploy-to-app (#5281)
# Motivation

When deploying NNS dapp to an app subnet, as we do for the Beta
canister, we upload assets separately to make sure the installed wasm
isn't too big.

Uploading the assets sometimes results in `"heap out of bounds"` when
done after an upgrade, as opposed to a reinstall.

We want to output canister logs after this process in case it helps
debug why this happens.

# Changes

Add `dfx canister logs` steps in the "deploy to app subnet" workflow.

# Tests

See example logs at
https://github.com/dfinity/nns-dapp/actions/runs/10113087618/job/27968618074

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary