Show detailed stake on neuron details page (#5296)
# Motivation

The neuron details page shows details for a single neuron. There's no
reason not to show the stake in full detail.
Also we want to round the stake in the table (in another PR) so it's
good to have the detailed stake displayed on the details page.

# Changes

1. Show detailed stake in the heading (for both NNS and SNS).
2. Show detailed stake in the stake action (which is shared between NNS
and SNS).

<img width="786" alt="Screenshot 2024-08-09 at 13 03 54"
src="https://github.com/user-attachments/assets/5e74d49d-879d-42b5-adc7-5a3b4e28156e">


# Tests

1. Unit test updated
2. Test manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/

# Todos

- [x] Add entry to changelog (if necessary).