Add universeId and tokenSymbol to TableProject (#5306)
# Motivation

We want to add a button on staking project rows for which the user
doesn't have any neurons yet, to allow the user to start staking those
tokens.

<img width="786" alt="image"
src="https://github.com/user-attachments/assets/f92bff2b-851c-4b1f-9eda-9f208a79d722">

We need to know the token symbol to render the button, and the universe
ID to open the staking modal when the button is clicked.

So this PR adds those 2 pieces of information to the `TableProject` type
to be used in the next PR.

# Changes

1. Add `universeId` and `tokenSymbol` fields to `TableProject`.
2. Fix `compareIcpFirst` to use the new `universeId` field instead of
depending on the `domKey`.

# Tests

Unit tests added.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet