Add Post-upgrade actions to release-sop (#5305)
# Motivation

The `release-sop` script has been very useful in making it easy to do
NNS dapp releases.
But it has been incomplete as it didn't include the post-upgrade steps.

# Changes

Add post upgrade steps to the script.

# Tests

Test json files were updated.
The latest release was made with these changes in the script.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary