fix: dogmi fee interpretation (#5314)
# Motivation

In the
[forum](https://forum.dfinity.org/t/dogmi-sns-transfer-wrong-transfer-fee/34114/1)
it was reported that the fees of the DOGMI SNS DAO were correctly
interpreted in the NNS dapp and Oisy. After debugging, we discovered
that the issue was linked to the SNS aggregator providing ICRC fees
encoded in Candid Nat, and our transformation to BigInt wasn't accurate.
It only interpreted the lower parts of a 64-bit value, i.e., only
interpreting the first 32 bits.

To resolve the issue, we implemented a new utility,
`candidNatArrayToBigInt`, which is included in `@dfinity/utils`.

# Notes

The upgrade of the utils that provides the new function was provided in
PR #5315.

# Changes

- Use new function to convert the SNS aggregator response to ICRC fee.
- We also use the new utility to convert potential `Value.Int`.
- Update aggregator types `Nat` and `Int` from single array number to a
potential `[number, number?]` type.

---------

Co-authored-by: Max Strasinsky <maksims.strasinskis@dfinity.org>
Co-authored-by: Max Strasinsky <98811342+mstrasinskis@users.noreply.github.com>