Explicitly upload checklist file in release-sop (#5334)
# Motivation

In the NNS dapp release SOP checklist there is a step:
> gh release upload "proposal-$PROPOSAL_ID"
scripts/nns-dapp/Release-*.json

To upload the release-sop JSON file to the release assets. Using `*`
makes it easy to copy/paste the command but in the `release-sop` script
we can be more precise and avoid matching other JSON.

When running the test there is a different JSON file but if an actual
release is ongoing the test would see that file too.

# Changes

1. Explicitly upload the checklist file for the current release instead
of using wild cards.
2. Drive-by: Fix typos of "relese" to "release".

# Tests

1. The existing `release-sop.test` still passes.
2. I tried running the test after doing `touch
scripts/nns-dapp/Release-2024-02-29.json` and it failed without the
change and passed with the change.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary