Rename labels to fix missing nnsFunction descriptions (#5336)
# Motivation

Some `nnsFunction` enum keys have been updated in
[ic-js](https://github.com/dfinity/ic-js/pull/695), but the
corresponding keys in the NNS-dapp governance labels were not updated.
This discrepancy leads to missing type field for the renamed types in
the NNS-dapp. In this PR, we synchronize the `nnsFunction` label keys to
match the `ic-js` enum.

# Changes

- Rename `UpdateSubnetReplicaVersion` to
`DeployGuestosToAllSubnetNodes`.
- Rename `UpdateElectedReplicaVersions` to
`ReviseElectedGuestosVersions`.
- Rename `AddApiBoundaryNode` to `AddApiBoundaryNodes`.
- Drive-by: update `getNnsFunctionKey` test name.

# Tests

- Added a test to verify that there are labels for all provided
`nnsFunction` values.
- Tested locally to confirm that the `DeployGuestosToAllSubnetNodes`
proposal now correctly displays the type field after the changes (which
was previously missing).

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.