NNS1-3250: Make project rows not clickable when neuron count not available (#5341)
# Motivation

We don't want users to navigate to the neurons table if they have no
neurons.
For this reason, we added the "Stake XYZ" button when the user has 0
neurons for a project.

But we also don't want to navigate if the user is not signed in or
neurons aren't loaded yet.

# Changes

1. When the number of neurons is not known, don't make the project row a
link.
2. When a non-link project row is clicked, only dispatch
`nnsStakeTokens` if the number of neurons is known to be 0.
3. Don't show the `>` icon when the row is neither a link nor clickable
to stake a neuron.

# Tests

1. Unit test updated and added.
2. Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/staking/

# Todos

- [x] Add entry to changelog (if necessary).