ci: add Andre and Arshavir as reviewers of automatic aggregator .did updates (#5365)
# Motivation

Since Arshavir and I are maintaining the SNS backend, and the aggregator
feels like it's more part of the sns backend than the frontend, it makes
sense that we review some of the PRs to the aggregator, and IIUC these
are the most common PRs. @dskloetd has been reviewing them until now, so
if we have questions for what to look for during the reviews, we can
maybe ask him for help.