NNS1-3281: Add update_locked_icp_e8s for TVL (#5373)
# Motivation

The NNS dapp displays the USD value of the total amount of ICP locked in
neurons.
It gets this information from the TVL canister but we want to move this
functionality to the nns-dapp canister and remove the TVL canister.

There are 2 inputs to this data:
1. The amount of ICP locked in neurons
2. The price of ICP in USD.

We plan to periodically load both values and cache them in the canister
state. When the user requests the TVL, it will be computed from the
currently cached state.

In this PR we only add a function which can load the amount of locked
ICP into the nns-dapp state.
The function is not yet called, and the state is not yet persisted in
this PR.

# Changes

1. Add `get_metrics` function to call the `get_metrics` function on the
governance canister. There is a separate implementation for testing.
2. Add a `TvlState` type which will hold all the state necessary to
compute the TVL. In this PR it will only have the `total_locked_icp_e8s`
for now.
3. Add the `TvlState` to the canister state. It's not yet included in
what is persisted to stable memory but should be in the future.
4. Add `update_locked_icp_e8s` which calls `get_metrics` and stores
`total_locked_icp_e8s` in the state.

# Tests

1. Added unit tests for the success and error cases.
2. Manually tested the feature end to end in a branch which has all the
other changes.


# Todos

- [ ] Add entry to changelog (if necessary).
Will add when the code is functional.